requirements.txt
python-dotenv
streamlit
langchain-huggingface
langchain_nvidia_ai_endpoints
langchain-community
faiss-cpu
pypdf